+++ dst/app/middlewares/authentication.js
@@ -10,33 +10,37 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.authentication = void 0;
-/**
- * oauthミドルウェア
- */
-const domain_1 = require("@chevre/domain");
-const express_middleware_1 = require("@motionpicture/express-middleware");
 const settings_1 = require("../settings");
-const authentication = (0, express_middleware_1.cognitoAuth)({
-    issuers: settings_1.TOKEN_ISSUERS,
-    authorizedHandler: (user, token, req, __, next) => __awaiter(void 0, void 0, void 0, function* () {
-        req.user = user;
-        req.accessToken = token;
-        next();
-    }),
-    unauthorizedHandler: (err, __1, __2, next) => {
-        // AbortErrorをハンドリング(2023-02-13~)
-        if (err.name === 'AbortError') {
-            next(new domain_1.chevre.factory.errors.ServiceUnavailable(`issuer unavailable. ${err.name}:${err.message}`));
-            // AmazonCognitoAPIのレート制限をハンドリング
+let cognitoAuthMiddeware;
+function authentication(req, res, next) {
+    return __awaiter(this, void 0, void 0, function* () {
+        if (cognitoAuthMiddeware === undefined) {
+            const expressMiddleware = yield Promise.resolve().then(() => require('@motionpicture/express-middleware'));
+            cognitoAuthMiddeware = expressMiddleware.cognitoAuth;
         }
-        else if (err.name === 'TooManyRequestsException') {
-            next(new domain_1.chevre.factory.errors.RateLimitExceeded(`getUser ${err.message}`));
-        }
-        else {
-            next(new domain_1.chevre.factory.errors.Unauthorized(`${err.name}:${err.message}`));
-        }
-    },
-    // タイムアウト設定(2023-04-17~)
-    requestOptions: { timeout: settings_1.TOKEN_ISSUER_REQUEST_TIMEOUT }
-});
+        return cognitoAuthMiddeware({
+            issuers: settings_1.TOKEN_ISSUERS,
+            authorizedHandler: (user, token, reqOnAuthorize, __, nextOnAuthorize) => __awaiter(this, void 0, void 0, function* () {
+                reqOnAuthorize.user = user;
+                reqOnAuthorize.accessToken = token;
+                nextOnAuthorize();
+            }),
+            unauthorizedHandler: (err, reqOnUnauthoize, __2, nextOnUnauthoize) => {
+                // AbortErrorをハンドリング(2023-02-13~)
+                if (err.name === 'AbortError') {
+                    nextOnUnauthoize(new reqOnUnauthoize.chevre.factory.errors.ServiceUnavailable(`issuer unavailable. ${err.name}:${err.message}`));
+                    // AmazonCognitoAPIのレート制限をハンドリング
+                }
+                else if (err.name === 'TooManyRequestsException') {
+                    nextOnUnauthoize(new reqOnUnauthoize.chevre.factory.errors.RateLimitExceeded(`getUser ${err.message}`));
+                }
+                else {
+                    nextOnUnauthoize(new reqOnUnauthoize.chevre.factory.errors.Unauthorized(`${err.name}:${err.message}`));
+                }
+            },
+            // タイムアウト設定(2023-04-17~)
+            requestOptions: { timeout: settings_1.TOKEN_ISSUER_REQUEST_TIMEOUT }
+        })(req, res, next);
+    });
+}
 exports.authentication = authentication;