+++ dst/app/middlewares/notFoundHandler.js
@@ -1,11 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.notFoundHandler = void 0;
-/**
- * 404ハンドラーミドルウェア
- */
-const domain_1 = require("@chevre/domain");
 function notFoundHandler(req, __, next) {
-    next(new domain_1.chevre.factory.errors.NotFound(`router for [${req.originalUrl}]`));
+    next(new req.chevre.factory.errors.NotFound(`router for [${req.originalUrl}]`));
 }
 exports.notFoundHandler = notFoundHandler;