+++ dst/app/middlewares/requireDomain.js
@@ -9,23 +9,21 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.validator = void 0;
-const express_validator_1 = require("express-validator");
-const http_status_1 = require("http-status");
-const api_1 = require("../error/api");
-function validator(req, __, next) {
+exports.requireDomain = void 0;
+let domain;
+function requireDomain(req, __, next) {
     return __awaiter(this, void 0, void 0, function* () {
-        const validatorResult = (0, express_validator_1.validationResult)(req);
-        if (!validatorResult.isEmpty()) {
-            const errors = validatorResult.array()
-                .map((mappedRrror) => {
-                return new req.chevre.factory.errors.Argument(mappedRrror.param, mappedRrror.msg);
-            });
-            next(new api_1.APIError(http_status_1.BAD_REQUEST, errors));
-        }
-        else {
+        try {
+            if (domain === undefined) {
+                const domainModule = yield Promise.resolve().then(() => require('@chevre/domain'));
+                domain = domainModule.chevre;
+            }
+            req.chevre = domain;
             next();
+        }
+        catch (error) {
+            next(error);
         }
     });
 }
-exports.validator = validator;
+exports.requireDomain = requireDomain;