+++ src/app/middlewares/authentication.ts
@@ -1,32 +1,38 @@
 /**
  * oauthミドルウェア
  */
-import { chevre } from '@chevre/domain';
 import { cognitoAuth } from '@motionpicture/express-middleware';
+import { NextFunction, Request, Response } from 'express';
 
 import { TOKEN_ISSUERS, TOKEN_ISSUER_REQUEST_TIMEOUT } from '../settings';
 
-const authentication = cognitoAuth({
-    issuers: TOKEN_ISSUERS,
-    authorizedHandler: async (user, token, req, __, next) => {
-        req.user = user;
-        req.accessToken = token;
+let cognitoAuthMiddeware: typeof cognitoAuth | undefined;
+export async function authentication(req: Request, res: Response, next: NextFunction) {
+    if (cognitoAuthMiddeware === undefined) {
+        const expressMiddleware = await import('@motionpicture/express-middleware');
+        cognitoAuthMiddeware = expressMiddleware.cognitoAuth;
+    }
 
-        next();
-    },
-    unauthorizedHandler: (err, __1, __2, next) => {
-        // AbortErrorをハンドリング(2023-02-13~)
-        if (err.name === 'AbortError') {
-            next(new chevre.factory.errors.ServiceUnavailable(`issuer unavailable. ${err.name}:${err.message}`));
-            // AmazonCognitoAPIのレート制限をハンドリング
-        } else if (err.name === 'TooManyRequestsException') {
-            next(new chevre.factory.errors.RateLimitExceeded(`getUser ${err.message}`));
-        } else {
-            next(new chevre.factory.errors.Unauthorized(`${err.name}:${err.message}`));
-        }
-    },
-    // タイムアウト設定(2023-04-17~)
-    requestOptions: { timeout: TOKEN_ISSUER_REQUEST_TIMEOUT }
-});
+    return cognitoAuthMiddeware({
+        issuers: TOKEN_ISSUERS,
+        authorizedHandler: async (user, token, reqOnAuthorize, __, nextOnAuthorize) => {
+            reqOnAuthorize.user = user;
+            reqOnAuthorize.accessToken = token;
 
-export { authentication };
+            nextOnAuthorize();
+        },
+        unauthorizedHandler: (err, reqOnUnauthoize, __2, nextOnUnauthoize) => {
+            // AbortErrorをハンドリング(2023-02-13~)
+            if (err.name === 'AbortError') {
+                nextOnUnauthoize(new reqOnUnauthoize.chevre.factory.errors.ServiceUnavailable(`issuer unavailable. ${err.name}:${err.message}`));
+                // AmazonCognitoAPIのレート制限をハンドリング
+            } else if (err.name === 'TooManyRequestsException') {
+                nextOnUnauthoize(new reqOnUnauthoize.chevre.factory.errors.RateLimitExceeded(`getUser ${err.message}`));
+            } else {
+                nextOnUnauthoize(new reqOnUnauthoize.chevre.factory.errors.Unauthorized(`${err.name}:${err.message}`));
+            }
+        },
+        // タイムアウト設定(2023-04-17~)
+        requestOptions: { timeout: TOKEN_ISSUER_REQUEST_TIMEOUT }
+    })(req, res, next);
+}