+++ src/app/middlewares/notFoundHandler.ts
@@ -1,9 +1,8 @@
 /**
  * 404ハンドラーミドルウェア
  */
-import { chevre } from '@chevre/domain';
 import { NextFunction, Request, Response } from 'express';
 
 export function notFoundHandler(req: Request, __: Response, next: NextFunction) {
-    next(new chevre.factory.errors.NotFound(`router for [${req.originalUrl}]`));
+    next(new req.chevre.factory.errors.NotFound(`router for [${req.originalUrl}]`));
 }