+++ src/app/routes/router.ts
@@ -9,10 +9,11 @@ import { ahRouter } from './_ah';
 import { accountTransactionsRouter } from './accountTransactions';
 import { cronRouter } from './cron';
 import { permitsRouter } from './permits';
-import { ssktsMembershipCouponRouter } from './ssktsMembershipCoupon';
-import { ssktsSurfrockRouter } from './ssktsSurfrock';
+// import { ssktsMembershipCouponRouter } from './ssktsMembershipCoupon';
+// import { ssktsSurfrockRouter } from './ssktsSurfrock';
 
 import { authentication } from '../middlewares/authentication';
+import { requireDomain } from '../middlewares/requireDomain';
 
 const router = express.Router();
 
@@ -29,15 +30,19 @@ router.get(
     }
 );
 router.use('/_ah', ahRouter);
-router.use('/cron', cronRouter);
 router.use('/health', healthRouter);
+
+// requireDomain(2023-10-06)
+router.use(requireDomain);
+
+router.use('/cron', cronRouter);
 
 // 認証
 router.use(authentication);
 
 router.use('/accountTransactions', accountTransactionsRouter);
 router.use('/permits', permitsRouter);
-router.use('/ssktsMembershipCoupon', ssktsMembershipCouponRouter);
-router.use('/ssktsSurfrock', ssktsSurfrockRouter);
+// router.use('/ssktsMembershipCoupon', ssktsMembershipCouponRouter); // 廃止(2023-10-06~)
+// router.use('/ssktsSurfrock', ssktsSurfrockRouter); // 廃止(2023-10-06~)
 
 export { router };